use hook instead
